+++ frontend/src/mocks/data.ts
@@ -5,8 +5,9 @@ import { CHARGER_TYPES, COMPANY_NAME } from '@constants/chargers';
 import { ENGLISH_DAYS } from '@constants/congestion';
 import { MAX_SEARCH_RESULTS } from '@constants/stationSearch';
 
-import type { CapacityType, ChargerDetails, CompanyName, MockStation } from '../types';
+import type { CompanyName, MockStation } from '../types';
 import type { ChargerType } from '../types';
+import type { CapacityType, ChargerDetails } from '../types/chargers';
 import type { Congestion, CongestionStatistics, EnglishDaysType } from '../types/congestion';
 
 const generateRandomData = <T>(array: T[]): T => {
